Creacion y Correcion de repository

+++ app/src/main/java/com/example/aplicacion/model/ProductoErrores.kt
@@ -3,7 +3,7 @@ package com.example.aplicacion.model
 data class ProductoErrores(
     val nombre: String? = null,
     val descripcion: String? = null,
-    val precio: Int? = null,
+    val precio: Double? = null,
     val urlImagen: String? = null,
     val categoria: String? = null
 )
