+++ app/src/main/java/com/example/aplicacion/model/ProductoUiState.kt
@@ -3,7 +3,7 @@ package com.example.aplicacion.model
 data class ProductoUiState(
     val nombre: String = "",
     val descripcion: String = "",
-    val precio: Int = 0,
+    val precio: Double = 0.0,
     val urlImagen: String = "",
     val categoria: String = "",
     val errores: ProductoErrores = ProductoErrores() //CONTENEDOR DE ERRORES
